training: match 8 model pipelines to OneTrainer/ai-toolkit + sampling-during-training
Fix campaign (build->skeptic->bugfix; all compile-clean, skeptic-verified vs reference):
- OneTrainer ref: klein(9b/4b), flux, zimage, ernie, sdxl, qwen.  ai-toolkit ref: hidream-o1, l2p.
- LR/warmup/alpha to per-model presets (OT default lora_alpha=1.0, scale=alpha/rank).
- qwen timestep discretization; ernie text-pad attention mask (253 garbage pad tokens were attending);
  zimage RoPE build_positions vs diffusers; sdxl alpha 16->1.0 + weight_decay + caption-dropout path;
  flux guidance 3.5->1.0 + kaiming-A + all-linear LoRA surface (504 modules vs OT empty layer_filter);
  hidream gauss-loss-weight off + sigma floor + 5 head adapters + save-key .model. strip;
  l2p real local_decoder head (killed proxy) + pixel/cap_feats cache reader + inverted-adaLN-timestep fix.
- Sampling-during-training wired: chroma/flux/anima/sdxl/qwen/sd35 (+ *_sample_resident modules).
- anima_sample_cli dtype fix (run-verified: produces a non-blank image).
Docs: TRAINING_PIPELINE_AUDIT + TRAINING_PIPELINE_FIXES (2026-06-22).
Caveat: compile + code-skeptic verified vs reference; real-run loss-trajectory parity is pending.

+++ serenitymojo/pipeline/anima_sample_cli.mojo
@@ -346,7 +346,7 @@ def _zero_b_set(set: AnimaLoraSet) -> AnimaLoraSet:
     for i in range(len(set.ad)):
         var src = set.ad[i].copy()
         for j in range(len(src.b)):
-            src.b[j] = Float32(0.0)
+            src.b[j] = BFloat16(0.0)
         ad.append(src^)
     return AnimaLoraSet(ad^, set.num_blocks, set.rank)
 
@@ -498,7 +498,7 @@ def _load_and_denoise(
         var babs = Float32(0.0)
         for i in range(len(lora.ad)):
             for j in range(len(lora.ad[i].b)):
-                var x = lora.ad[i].b[j]
+                var x = Float32(lora.ad[i].b[j])
                 babs += x if x >= 0.0 else -x
         print("  LoRA overlay loaded:", len(lora.ad), "adapters  |B|_1 =", babs,
               " scale=alpha/rank=", ALPHA / Float32(RANK))
